requirements.txt
duckduckgo-search>=6.2.13
requests>=2.31.0
beautifulsoup4>=4.12.2
httpx>=0.27.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
tqdm>=4.66.4
pydantic>=2.7.1
openai>=1.30.1
tenacity>=8.4.2
lxml>=5.2.2